refactor: share prometheus formatting primitives via common-env
