+++ packages/common-env/src/index.ts
@@ -148,5 +148,6 @@ export function quoteShellArgv(argv: string[]): string {
 
 export * from "./deployment-settings";
 export * from "./metrics-auth";
+export * from "./metrics-format";
 export * from "./service-url";
 export * from "./tether-protocol";
